refactor: global var collection to utils
